@@ -1,5 +1,6 @@
 ﻿using BenchmarkDotNet.Running;
 using System;
+using System.Runtime.InteropServices;
 
 namespace Benchmark
 {
@@ -7,11 +8,9 @@ namespace Benchmark
     {
         private static void Main(string[] args)
         {
-            if (args == null || args.Length == 0)
+            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows) && (args is null || args.Length == 0))
             {   // if no args, we're probably using Ctrl+F5 in the IDE; enlargen thyself!
-#pragma warning disable CA1416 // windows only
                 try { Console.WindowWidth = Console.LargestWindowWidth - 20; } catch { }
-#pragma warning restore CA1416 // windows only
             }
             BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
         }
