@@ -21,29 +21,29 @@ namespace Pipelines.Sockets.Unofficial.Tests
         }
 
         private ITestOutputHelper Output { get; }
-        private TestTextWriter Log { get; }
 
         public ConnectTests(ITestOutputHelper output)
         {
             Output = output;
-            Log = TestTextWriter.Create(output);
         }
 
         [Fact]
         public async Task Connect()
         {
-            var timeout = Task.Delay(5000);
+            var timeout = Task.Delay(6000);
             var code = ConnectImpl();
             var first = await Task.WhenAny(timeout, code).ConfigureAwait(false);
-            if (first == timeout) Throw.Timeout();
-            await first;
+            if (first == timeout) Throw.Timeout("unknown timeout");
+            await first; // check outcome
         }
 
         private async Task ConnectImpl()
         {
-            var endpoint = new IPEndPoint(IPAddress.Loopback, 9080);
+            int port = 16320 + new Random().Next(100);
+            var endpoint = new IPEndPoint(IPAddress.Loopback, port);
             object waitForRunning = new object();
             Task<string> server;
+            Output.WriteLine("Starting server...");
             lock (waitForRunning)
             {
                 server = Task.Run(() => SyncEchoServer(waitForRunning, endpoint));
@@ -51,81 +51,96 @@ namespace Pipelines.Sockets.Unofficial.Tests
                     Throw.Timeout("Server didn't start");
             }
 
+            if (server.IsFaulted)
+            {
+                await server; // early exit if broken
+            }
+
             string actual;
-            Log?.DebugLog("connecting...");
+            Output.WriteLine("connecting...");
             using var conn = await SocketConnection.ConnectAsync(endpoint,
                 connectionOptions: SocketConnectionOptions.ZeroLengthReads).ConfigureAwait(false);
             var data = Encoding.ASCII.GetBytes("Hello, world!");
-            Log?.DebugLog("sending message...");
+            Output.WriteLine("sending message...");
             await conn.Output.WriteAsync(data).ConfigureAwait(false);
-
-#if NET6_0_OR_GREATER
+          
             Assert.True(conn.Output.CanGetUnflushedBytes, "conn.Output.CanGetUnflushedBytes");
-#endif
 
-            Log?.DebugLog("completing output");
+            Output.WriteLine("completing output");
             conn.Output.Complete();
 
-            Log?.DebugLog("awaiting server...");
+            Output.WriteLine("awaiting server...");
             actual = await server;
 
             Assert.Equal("Hello, world!", actual);
 
             string returned;
-            Log?.DebugLog("buffering response...");
+            Output.WriteLine("buffering response...");
             while (true)
             {
                 var result = await conn.Input.ReadAsync().ConfigureAwait(false);
 
                 var buffer = result.Buffer;
-                Log?.DebugLog($"received {buffer.Length} bytes");
+                Output.WriteLine($"received {buffer.Length} bytes");
                 if (result.IsCompleted)
                 {
                     returned = Encoding.ASCII.GetString(result.Buffer.ToArray());
-                    Log?.DebugLog($"received: '{returned}'");
+                    Output.WriteLine($"received: '{returned}'");
                     break;
                 }
 
-                Log?.DebugLog("advancing");
+                Output.WriteLine("advancing");
                 conn.Input.AdvanceTo(buffer.Start, buffer.End);
             }
 
             Assert.Equal("!dlrow ,olleH", returned);
 
-            Log?.DebugLog("disposing");
+            Output.WriteLine("disposing");
         }
 
         private Task<string> SyncEchoServer(object ready, IPEndPoint endpoint)
         {
-            var listener = new TcpListener(endpoint);
-            Log?.DebugLog($"[Server] starting on {endpoint}...");
-            listener.Start();
-            Output.WriteLine("[Server] running; waiting for connection...");
-            lock (ready)
+            try
             {
-                Monitor.Pulse(ready);
-            }
-            string s;
-            using (var socket = listener.AcceptSocket())
-            {
-                Output.WriteLine($"[Server] accepted connection");
-                using var ns = new NetworkStream(socket);
-                using (var reader = new StreamReader(ns, Encoding.ASCII, false, 1024, true))
-                using (var writer = new StreamWriter(ns, Encoding.ASCII, 1024, true))
+                var listener = new TcpListener(endpoint);
+                Output.WriteLine($"[Server] starting on {endpoint}...");
+                listener.Start();
+                lock (ready)
                 {
-                    s = reader.ReadToEnd();
-                    Output.WriteLine($"[Server] received '{s}'; replying in reverse...");
-                    char[] chars = s.ToCharArray();
-                    Array.Reverse(chars);
-                    var t = new string(chars);
-                    writer.Write(t);
+                    Monitor.PulseAll(ready);
                 }
-                socket.Shutdown(SocketShutdown.Both);
-                socket.Close();
+                Output.WriteLine("[Server] running; waiting for connection...");
+                string s;
+                using (var socket = listener.AcceptSocket())
+                {
+                    Output.WriteLine($"[Server] accepted connection");
+                    using var ns = new NetworkStream(socket);
+                    using (var reader = new StreamReader(ns, Encoding.ASCII, false, 1024, true))
+                    using (var writer = new StreamWriter(ns, Encoding.ASCII, 1024, true))
+                    {
+                        s = reader.ReadToEnd();
+                        Output.WriteLine($"[Server] received '{s}'; replying in reverse...");
+                        char[] chars = s.ToCharArray();
+                        Array.Reverse(chars);
+                        var t = new string(chars);
+                        writer.Write(t);
+                    }
+                    socket.Shutdown(SocketShutdown.Both);
+                    socket.Close();
+                }
+                Output.WriteLine($"[Server] shutting down");
+                listener.Stop();
+                return Task.FromResult(s);
             }
-            Output.WriteLine($"[Server] shutting down");
-            listener.Stop();
-            return Task.FromResult(s);
+            catch (Exception ex)
+            {
+                Output.WriteLine($"[Server] faulted: {ex.Message}");
+                lock (ready)
+                {
+                    Monitor.PulseAll(ready);
+                }
+                return Task.FromException<string>(ex);
+            }
         }
     }
 }
